PI Estimator Tool for Sehir CS projects.
The quality of an academic institution can be evaluated according to the research projects developed by the academicians and the funds received by these projects. The Department of Computer Science at Istanbul Sehir University is home to many research projects developed by academicians and funded by authorized institutions. In this project, the main goal is, building a tool that automatically predicts the principal investigators of research projects which are carried out by Sehir CS Faculty Members.